delete comments clean files
