+++ lib/enigma.rb
@@ -1,7 +1,5 @@
-require_relative "makeable"
 require_relative 'ciphen'
 class Enigma < Ciphen
-  include Makeable
   attr_reader :encrypt, :decrypt
   def encrypt(message, key = make_key, date = make_date)
     encrypted_msg = encrypt_msg(message, key, date)
@@ -20,5 +18,4 @@ class Enigma < Ciphen
       date:       date
       })
   end
-
 end
